+++ src/av/globals.c
@@ -1,0 +1,6 @@
+#include "video.h"
+
+/* Globals */
+
+/* video.h */
+uint16_t video_width, video_height, max_video_width, max_video_height;
