+++ src/theme_tables.h
@@ -3,153 +3,7 @@
 
 #include "theme.h"
 
-const char *COLOUR_NAME_TABLE[] = { "MAIN_BACKGROUND",
-                                    "ALT_BACKGROUND",
-                                    "MAIN_TEXT",
-                                    "MAIN_CHATTEXT",
-                                    "MAIN_SUBTEXT",
-                                    "MAIN_ACTIONTEXT",
-                                    "MAIN_QUOTETEXT",
-                                    "MAIN_REDTEXT",
-                                    "MAIN_URLTEXT",
-                                    "MAIN_HINTTEXT",
-
-                                    "MENU_BACKGROUND",
-                                    "MENU_TEXT",
-                                    "MENU_SUBTEXT",
-                                    "MENU_BKGRND_HOVER",
-                                    "MENU_ACTIVE_BACKGROUND",
-                                    "MENU_ACTIVE_TEXT",
-
-                                    "MSG_USER",
-                                    "MSG_USER_PEND",
-                                    "MSG_USER_ERROR",
-                                    "MSG_CONTACT",
-
-                                    "LIST_BACKGROUND",
-                                    "LIST_BKGRND_HOVER",
-                                    "LIST_TEXT",
-                                    "LIST_SUBTEXT",
-
-                                    "GROUP_SELF",
-                                    "GROUP_PEER",
-                                    "GROUP_AUDIO",
-                                    "GROUP_MUTED",
-
-                                    "SELECTION_BACKGROUND",
-                                    "SELECTION_TEXT",
-
-                                    "EDGE_NORMAL",
-                                    "EDGE_HOVER",
-                                    "EDGE_ACTIVE",
-                                    "ACTIVEOPTION_BACKGROUND",
-                                    "ACTIVEOPTION_TEXT",
-
-                                    "AUX_BACKGROUND",
-                                    "AUX_EDGE_NORMAL",
-                                    "AUX_EDGE_HOVER",
-                                    "AUX_EDGE_ACTIVE",
-                                    "AUX_TEXT",
-                                    "AUX_ACTIVEOPTION_BACKGROUND",
-                                    "AUX_ACTIVEOPTION_TEXT",
-
-                                    "STATUS_ONLINE",
-                                    "STATUS_AWAY",
-                                    "STATUS_BUSY",
-
-                                    "BUTTON_SUCCESS_BACKGROUND",
-                                    "BUTTON_SUCCESS_TEXT",
-                                    "BUTTON_SUCCESS_BKGRND_HOVER",
-                                    "BUTTON_SUCCESS_TEXT_HOVER",
-                                    "BUTTON_WARNING_BACKGROUND",
-                                    "BUTTON_WARNING_TEXT",
-                                    "BUTTON_WARNING_BKGRND_HOVER",
-                                    "BUTTON_WARNING_TEXT_HOVER",
-                                    "BUTTON_DANGER_BACKGROUND",
-                                    "BUTTON_DANGER_TEXT",
-                                    "BUTTON_DANGER_BKGRND_HOVER",
-                                    "BUTTON_DANGER_TEXT_HOVER",
-                                    "BUTTON_DISABLED_BACKGROUND",
-                                    "TRANSFER_PROGRESS_OVERLAY_PAUSED",
-                                    "BUTTON_DISABLED_TEXT",
-                                    "BUTTON_DISABLED_TRANSFER",
-                                    "BUTTON_INPROGRESS_BACKGROUND",
-                                    "TRANSFER_PROGRESS_OVERLAY_ACTIVE",
-                                    "BUTTON_INPROGRESS_TEXT",
-                                    NULL };
-uint32_t *COLOUR_POINTER_TABLE[] = { &COLOR_BKGRND_MAIN,
-                                     &COLOR_BKGRND_ALT,
-                                     &COLOR_MAIN_TEXT,
-                                     &COLOR_MAIN_TEXT_CHAT,
-                                     &COLOR_MAIN_TEXT_SUBTEXT,
-                                     &COLOR_MAIN_TEXT_ACTION,
-                                     &COLOR_MAIN_TEXT_QUOTE,
-                                     &COLOR_MAIN_TEXT_RED,
-                                     &COLOR_MAIN_TEXT_URL,
-                                     &COLOR_MAIN_TEXT_HINT,
-
-                                     &COLOR_BKGRND_MENU,
-                                     &COLOR_MENU_TEXT,
-                                     &COLOR_MENU_TEXT_SUBTEXT,
-                                     &COLOR_BKGRND_MENU_HOVER,
-                                     &COLOR_BKGRND_MENU_ACTIVE,
-                                     &COLOR_MENU_TEXT_ACTIVE,
-
-                                     &COLOR_MSG_USER,
-                                     &COLOR_MSG_USER_PEND,
-                                     &COLOR_MSG_USER_ERROR,
-                                     &COLOR_MSG_CONTACT,
-
-                                     &COLOR_BKGRND_LIST,
-                                     &COLOR_BKGRND_LIST_HOVER,
-                                     &COLOR_LIST_TEXT,
-                                     &COLOR_LIST_TEXT_SUBTEXT,
-
-                                     &COLOR_GROUP_SELF,
-                                     &COLOR_GROUP_PEER,
-                                     &COLOR_GROUP_AUDIO,
-                                     &COLOR_GROUP_MUTED,
-
-                                     &COLOR_SELECTION_BACKGROUND,
-                                     &COLOR_SELECTION_TEXT,
-
-                                     &COLOR_EDGE_NORMAL,
-                                     &COLOR_EDGE_HOVER,
-                                     &COLOR_EDGE_ACTIVE,
-                                     &COLOR_ACTIVEOPTION_BKGRND,
-                                     &COLOR_ACTIVEOPTION_TEXT,
-
-                                     &COLOR_BKGRND_AUX,
-                                     &COLOR_AUX_EDGE_NORMAL,
-                                     &COLOR_AUX_EDGE_HOVER,
-                                     &COLOR_AUX_EDGE_ACTIVE,
-                                     &COLOR_AUX_TEXT,
-                                     &COLOR_AUX_ACTIVEOPTION_BKGRND,
-                                     &COLOR_AUX_ACTIVEOPTION_TEXT,
-
-                                     &COLOR_STATUS_ONLINE,
-                                     &COLOR_STATUS_AWAY,
-                                     &COLOR_STATUS_BUSY,
-
-                                     &COLOR_BTN_SUCCESS_BKGRND,
-                                     &COLOR_BTN_SUCCESS_TEXT,
-                                     &COLOR_BTN_SUCCESS_BKGRND_HOVER,
-                                     &COLOR_BTN_SUCCESS_TEXT_HOVER,
-                                     &COLOR_BTN_WARNING_BKGRND,
-                                     &COLOR_BTN_WARNING_TEXT,
-                                     &COLOR_BTN_WARNING_BKGRND_HOVER,
-                                     &COLOR_BTN_WARNING_TEXT_HOVER,
-                                     &COLOR_BTN_DANGER_BACKGROUND,
-                                     &COLOR_BTN_DANGER_TEXT,
-                                     &COLOR_BTN_DANGER_BKGRND_HOVER,
-                                     &COLOR_BTN_DANGER_TEXT_HOVER,
-                                     &COLOR_BTN_DISABLED_BKGRND,
-                                     &COLOR_BTN_DISABLED_FORGRND,
-                                     &COLOR_BTN_DISABLED_TEXT,
-                                     &COLOR_BTN_DISABLED_TRANSFER,
-                                     &COLOR_BTN_INPROGRESS_BKGRND,
-                                     &COLOR_BTN_INPROGRESS_FORGRND,
-                                     &COLOR_BTN_INPROGRESS_TEXT,
-                                     NULL };
+extern const char *COLOUR_NAME_TABLE[];
+extern uint32_t *COLOUR_POINTER_TABLE[];
 
 #endif
