+++ src/ui/globals.c
@@ -1,0 +1,6 @@
+#include "draw.h"
+
+/* Globals */
+
+/* draw.h */
+int font_small_lineheight, font_msg_lineheight;
